refactor(Combat): refactored Combat. Clean and less buggy
