+++ src/views/Combat/components/CombatSetup/SetupToolbar.tsx
@@ -1,4 +1,6 @@
-import { Button, Tooltip, Grid, Select, MenuItem, InputLabel, useTheme } from "@mui/material";
+import { useState } from "react";
+import { Button, Tooltip, Grid, Select, MenuItem, InputLabel, useTheme, SelectChangeEvent } from "@mui/material";
+import ConfirmationDialog from "../Dialogs/ConfirmationDialog";
 
 export default function CombatSetupToolbar({
   onStartCombat,
@@ -6,12 +8,13 @@ export default function CombatSetupToolbar({
   scenarioName,
   onAddActor,
   onDeleteActor,
-  isValidActors,
+  isValidSetup,
 }) {
   const theme = useTheme();
+  const [dialogOpen, setDialogOpen] = useState(false);
 
   return (
-    <Grid item container>
+    <>
       <Grid item container spacing={4} alignItems="center">
         <Grid item xs>
           <InputLabel id="select-scenario-label" sx={{ color: "white" }}>
@@ -21,7 +24,7 @@ export default function CombatSetupToolbar({
             labelId="select-scenario-label"
             label="Scenario"
             value={scenarioName}
-            onChange={onScenarioChange}
+            onChange={(e: SelectChangeEvent<any>) => onScenarioChange(e.target.value)}
             style={{ width: "100%", color: "white", backgroundColor: `${theme.palette.secondary.main}` }}
           >
             <MenuItem value="skeletons">Skeletons</MenuItem>
@@ -30,7 +33,7 @@ export default function CombatSetupToolbar({
           </Select>
         </Grid>
         <Grid xs sx={{ padding: "55px 10px 0px 20px" }}>
-          <Button variant="contained" fullWidth onClick={onAddActor}>
+          <Button variant="contained" fullWidth onClick={() => setDialogOpen(true)}>
             Add Actor
           </Button>
         </Grid>
@@ -40,7 +43,7 @@ export default function CombatSetupToolbar({
           </Button>
         </Grid>
         <Grid xs sx={{ padding: "55px 0px 0px 10px" }}>
-          {isValidActors ? (
+          {isValidSetup ? (
             <Button variant="contained" fullWidth onClick={onStartCombat}>
               Start Combat
             </Button>
@@ -55,6 +58,7 @@ export default function CombatSetupToolbar({
           )}
         </Grid>
       </Grid>
-    </Grid>
+      <ConfirmationDialog open={dialogOpen} setOpen={setDialogOpen} onClose={onAddActor} />
+    </>
   );
 }
