+++ src/views/Combat/components/CombatStart/CombatStart.tsx
@@ -1,0 +1,8 @@
+import { useState } from "react";
+import { GridRowParams, GridCellEditCommitParams } from "@mui/x-data-grid";
+import CombatToolbar from "./CombatToolbar";
+import CombatDataTable from "../../../../components/DataTables/CombatDataTable";
+
+export default function CombatStart() {
+  return <div>Working on it</div>;
+}
